git-reparent: Tweak the syntax for specifying a commit range
Git already uses `A..B` as syntax for specifying commit ranges.
However, unlike in Mercurial where `A..B` specifies an inclusive
range like we want, `A..B` in Git is a half-open interval where
where `A` is an exclusive endpoint.

To avoid confusion with normal Git syntax, borrow Hg's alternative
`A::B` syntax for specifying an inclusive commit range. (`:` is not
a legal character for Git refnames.)