git-resolve: Always prompt to delete files for deletion conflicts
When resolving a deletion conflict, regardless of whether the
file was deleted by "us", by "them", or by both, usually we'd
want the file to stay deleted.  Previously `git-resolve`
prompted to automatically delete the file only for the "both"
case, but prompt for the other cases too.

Also rename soeme class members to indicate whether they
should be treated as private members.